+++ scripts/customer_contract_details.py
@@ -568,40 +568,73 @@ def customer_contract_details(master_files, path):
     # Module Group shipping frequency must match with Shipping Calendar frequency
     # Note: This solution iterates to the last date range
     def customer_contract_details_ttc_contract_2(cell_row, cell_col, new_mod):
-        module_shipping_frequency = {}
+        module_group_code = master_files['xl_sheet_main'].cell_value(cell_row, cell_col-2)
+        ttc_contract_no = master_files['xl_sheet_main'].cell_value(cell_row, cell_col)
+
+        module_shipping_frequency = ''
+        for row in range(10, selected['backup_5'].sheet_by_index(0).nrows):
+            if selected['backup_5'].sheet_by_index(0).cell_value(row, 2) == module_group_code:
+                module_shipping_frequency = selected['backup_5'].sheet_by_index(0).cell_value(row, 8)
+
         try:
-            for row in range(10, additional['TNM_MODULE_GROUP'].nrows):
-                module_shipping_frequency[additional['TNM_MODULE_GROUP'].cell_value(row, 2)] = additional['TNM_MODULE_GROUP'].cell_value(row, 8)
+            for row in range(9, additional['TNM_MODULE_GROUP'].nrows):
+                # If in submitted, regardless of NEW or MOD, should replace backup
+                if additional['TNM_MODULE_GROUP'].cell_value(row, 2) == module_group_code:
+                    module_shipping_frequency = additional['TNM_MODULE_GROUP'].cell_value(row, 8)
         except KeyError:
             pass
-        finally:
-            for row in range(10, selected['backup_5'].sheet_by_index(0).nrows):
-                module_shipping_frequency[selected['backup_5'].sheet_by_index(0).cell_value(row, 2)] = selected['backup_5'].sheet_by_index(0).cell_value(row, 8)
 
-        shipping_calendar = {}
-        for row in range(9, selected['backup_8'].sheet_by_index(0).nrows):
-            shipping_calendar[selected['backup_8'].sheet_by_index(0).cell_value(row, 2)] = selected['backup_8'].sheet_by_index(0).cell_value(row, 12)
+        if module_shipping_frequency == '':
+            print ('TTC Contract No. check 2 --- Fail')
+            update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'FAIL', ttc_contract_no, module_group_code, 'Module Group Code cannot be found in Module Group Master')
+            return
 
-        ttc_contract_route = {}
+        shipping_route = ''
+        for row in range(9, selected['backup_4'].sheet_by_index(0).nrows):
+            if selected['backup_4'].sheet_by_index(0).cell_value(row, 2) == ttc_contract_no:
+                shipping_route = selected['backup_4'].sheet_by_index(0).cell_value(row, 14)
+
         try:
             for row in range(9, additional['TNM_TTC_CONTRACT'].nrows):
-                ttc_contract_route[additional['TNM_TTC_CONTRACT'].cell_value(row, 2)] = additional['TNM_TTC_CONTRACT'].cell_value(row, 14)
+                # If in submitted, regardless of NEW or MOD, should replace backup
+                if additional['TNM_TTC_CONTRACT'].cell_value(row, 2) == ttc_contract_no:
+                    shipping_route = additional['TNM_TTC_CONTRACT'].cell_value(row, 14)
         except KeyError:
             pass
-        finally:
-            for row in range(9, selected['backup_4'].sheet_by_index(0).nrows):
-                ttc_contract_route[selected['backup_4'].sheet_by_index(0).cell_value(row, 2)] = selected['backup_4'].sheet_by_index(0).cell_value(row, 14)
+
+        if shipping_route == '':
+            print ('Shipping Frequency check --- Fail')
+            update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'FAIL', ttc_contract_no, ttc_contract_no, 'TTC Contract cannot be found in TTC Contract Master')
+            return
+
+        shipping_calendar_frequency = ''
+        for row in range(9, selected['backup_8'].sheet_by_index(0).nrows):
+            # Loop will end with last ETD
+            if selected['backup_8'].sheet_by_index(0).cell_value(row, 2) == shipping_route:
+                shipping_calendar_frequency = selected['backup_8'].sheet_by_index(0).cell_value(row, 12)
+
+        if shipping_calendar_frequency == '':
+            print ('Shipping Frequency check --- Fail')
+            update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'FAIL', ttc_contract_no, shipping_route, 'Shipping Route cannot be found in Shipping Calendar Master')
+            return
+
+        # Split frequency into elements, convert to set
+        try:
+            shipping_calendar_frequency_split = set(shipping_calendar_frequency.split(','))
+        except AttributeError:
+            shipping_calendar_frequency_split = {int(shipping_calendar_frequency)}
 
         try:
-            if module_shipping_frequency[master_files['xl_sheet_main'].cell_value(cell_row, cell_col-2)] == shipping_calendar[ttc_contract_route[master_files['xl_sheet_main'].cell_value(cell_row, cell_col)]]:
-                print ('TTC Contract No. check 2 --- Pass (%s)' % master_files['xl_sheet_main'].cell_value(cell_row, cell_col))
-                update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'PASS', shipping_calendar[ttc_contract_route[master_files['xl_sheet_main'].cell_value(cell_row, cell_col)]], module_shipping_frequency[master_files['xl_sheet_main'].cell_value(cell_row, cell_col-2)], 'Module Group shipping frequency match with shipping frequency of last ETD of Shipping Route')
-            else:
-                print ('TTC Contract No. check 2 --- Fail (Shipping Frequencies do not match)')
-                update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'WARNING', shipping_calendar[ttc_contract_route[master_files['xl_sheet_main'].cell_value(cell_row, cell_col)]], module_shipping_frequency[master_files['xl_sheet_main'].cell_value(cell_row, cell_col-2)], 'Module Group shipping frequency does not match with shipping frequency of last ETD of Shipping Route, please check if date overlaps with other ETDs')
-        except KeyError:
-            print ('TTC Contract No. check 2 --- Fail (Module Group / TTC Contract not found)')
-            update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'FAIL', master_files['xl_sheet_main'].cell_value(cell_row, cell_col), 'NA', 'Module Group Code / TTC Contract No. not found in system or submited masters')
+            module_shipping_frequency_split = set(module_shipping_frequency.split(','))
+        except AttributeError:
+            module_shipping_frequency_split = {str(int(module_shipping_frequency))}
+
+        if len(set.intersection(shipping_calendar_frequency_split, module_shipping_frequency_split)) != 0:
+            print ('Shipping Frequency check --- Pass')
+            update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'PASS', module_shipping_frequency, shipping_calendar_frequency, 'Module Group shipping frequency intersects with shipping frequency of last ETD of Shipping Route')
+        else:
+            print ('Shipping Frequency check --- Fail')
+            update_df(new_mod, columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'FAIL', module_shipping_frequency, shipping_calendar_frequency, 'Module Group shipping frequency does not intersect with shipping frequency of last ETD of Shipping Route, please check if date overlaps with other ETDs')
 
     # All parts in 1 Module Group must have the same TTC Contract No
     def customer_contract_details_ttc_contract_3(cell_row, cell_col, new_mod):